+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5 pre
 * Digest 增加 InputStream 系列方法
 * Collectionx 增加 joinToString(@NotNull Iterable<T>, @Nullable CharSequence, @NotNull Transformer<T, CharSequence>) 方法
+* File mkdirsWithThrow() 重命名为 mkdirsOrThrow，createNewFileWithThrow 重命名为 createNewFileOrThrow
 
 ## v0.4